requirements_dev.txt
tox==2.1.1
docopt>=0.6.2
PyYaml>=3.11
python-dateutil>=2.4.2
pykwalify